修复了下拉刷新精品歌单不变的bug 但是性能的问题还是没有得到解决

+++ app/src/main/java/com/example/neteasecloudmusic/favoriteslist/FavoritesModel.kt
@@ -187,14 +187,14 @@ class FavoritesModel :FavoritesContract.FavoritesIModel{
         var br: Int = 0,
         var fid: Int = 0,
         var size: Int = 0,
-        var vd: Int = 0
+        var vd: Double = 0.0
     )
 
     data class M(
         var br: Int = 0,
         var fid: Int = 0,
         var size: Int = 0,
-        var vd: Int = 0
+        var vd: Double = 0.0
     )
 
     data class ChargeInfo(
